+++ {{ cookiecutter.project_name }}/project/static-sources/ckeditor/ckeditor/plugins/vimeo/lang/en.js
@@ -1,0 +1,18 @@
+CKEDITOR.plugins.setLang('vimeo', 'en', {
+    button : 'Embed Vimeo Video',
+    title : 'Embed Vimeo Video',
+    txtEmbed : 'Paste Embed Code Here',
+    txtUrl : 'Paste Vimeo Video URL',
+    txtWidth : 'Width',
+    txtHeight : 'Height',
+    chkAutoplay : 'Autoplay',
+    noCode : 'You must input an embed code or URL',
+    invalidEmbed : 'The embed code you\'ve entered doesn\'t appear to be valid',
+    invalidUrl : 'The URL you\'ve entered doesn\'t appear to be valid',
+    or : 'or',
+    noWidth : 'You must inform the width',
+    invalidWidth : 'Inform a valid width',
+    noHeight : 'You must inform the height',
+    invalidHeight : 'Inform a valid height',
+    txtResponsive : 'Make Responsive (ignore width and height, fit to width)'
+});