+++ {{ cookiecutter.project_name }}/project/static-sources/ckeditor/ckeditor/plugins/vimeo/lang/fr.js
@@ -1,0 +1,18 @@
+CKEDITOR.plugins.setLang('vimeo', 'fr', {
+    button : 'Insérer une vidéo Vimeo',
+    title : 'Insérer une vidéo Vimeo',
+    txtEmbed : 'Coller le code embed ici',
+    txtUrl : 'Coller l\'url de la vidéo ici',
+    txtWidth : 'Largeur',
+    txtHeight : 'Hauteur',
+    chkAutoplay : 'Autoplay',
+    noCode : 'Vous devez entrer un code embed ou une url',
+    invalidEmbed : 'Le code embed est invalide',
+    invalidUrl : 'L\'url est invalide',
+    or : 'ou',
+    noWidth : 'Vous devez saisir une largeur',
+    invalidWidth : 'La largeur saisie est invalide',
+    noHeight : 'Vous devez saisir une hauteur',
+    invalidHeight : 'La hauteur saisie est invalide',
+    txtResponsive : 'Responsive video'
+});